+++ bioprov/programs/programs.py
@@ -2,13 +2,15 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.18a"
+__version__ = "0.1.19"
 
 
 """
 Module for holding preset instances of the Program class.
+Module for holding preset instances of the Program class.
 """
 
+import logging
 from os import path
 from pathlib import Path
 
@@ -117,8 +119,8 @@ def blastp(sample, db, query_tag="query", outformat=6):
     :param Sample sample: Instance of BioProv.Sample.
     :param str db: A string pointing to the reference database directory and title.
     :param str query_tag: A tag for the query file.
-    :param int outformat: The output format to gather from blastn.
-    :return: Instance of PresetProgram for BLASTN.
+    :param int outformat: The output format to gather from blastp.
+    :return: Instance of PresetProgram for BLASTP.
     :rtype: BioProv.PresetProgram.
     :raises AssertionError: Path to the reference database does not exist.
     """
@@ -192,10 +194,12 @@ def prokka(
     )
 
     for param in params:
-        _prokka.add_parameter(param, _print=False)
+        _prokka.add_parameter(param)
 
     if path.isdir(output_path):
-        print(f"Warning: {output_path} directory exists. Will overwrite.")  # no cover
+        config.logger.warning(
+            f"Warning: {output_path} directory exists. Will overwrite."
+        )  # no cover
         _prokka.add_parameter(
             Parameter(key="--force", value="", kind="misc")
         )  # no cover
@@ -223,8 +227,7 @@ def prokka(
 
     # Input goes here, must be last positionally.
     _prokka.add_parameter(
-        Parameter(key="", value=str(_sample.files[assembly]), kind="input"),
-        _print=False,
+        Parameter(key="", value=str(_sample.files[assembly]), kind="input")
     )
 
     return _prokka
@@ -274,7 +277,7 @@ def kaiju(
         Parameter(key="-o", value=output_path, kind="output"),
     )
     for p in params:
-        kaiju_.add_parameter(p, _print=False)
+        kaiju_.add_parameter(p)
 
     if add_param_str:
         kaiju_.cmd += f" {add_param_str}"  # no cover
@@ -324,7 +327,7 @@ def kaiju2table(
     )
 
     for p in params:
-        kaiju2table_.add_parameter(p, _print=False)
+        kaiju2table_.add_parameter(p)
 
     # Add final parameter:
     kaiju2table_.cmd += f" {str(_sample.files[kaiju_output])}"