+++ bioprov/src/files.py
@@ -2,17 +2,20 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.18a"
+__version__ = "0.1.19"
 
 
 """
 Contains the File and SeqFile classes and related functions.
 """
 
-import numpy as np
+import logging
 from dataclasses import dataclass
 from pathlib import Path
+
+import numpy as np
 from Bio import SeqIO, AlignIO
+
 from bioprov.utils import (
     get_size,
     Warnings,
@@ -113,12 +116,12 @@ class File:
         # TODO: replace these print statements for logger warning/debug level
         if warnings:
             if not self.exists and old_exists:
-                print(
-                    f"Warning: file {self.path} was marked as existing but was not found."
+                logging.warning(
+                    f"File {self.path} was marked as existing but was not found."
                 )
             if old_hash and self.sha1 != old_hash and self.exists:  # no cover
-                print(
-                    f"Warning: file {self.path} previous sha1 checksum differs from the current."
+                logging.warning(
+                    f"File {self.path} previous sha1 checksum differs from the current."
                 )
 
     def serializer(self):
@@ -159,7 +162,7 @@ class Directory:
         # TODO: replace these print statements for logger warning/debug level
         if warnings:
             if not self.exists and old_exists:
-                print(
+                logging.warning(
                     f"Warning: file {self.path} was marked as existing but was not found."
                 )
 
@@ -448,8 +451,8 @@ def seqrecordgenerator(path, format, parser="seq", warnings=False):
         return records
     except FileNotFoundError:
         if warnings:
-            print(Warnings()["not_exist"](path))
-            print(
+            logging.warning(Warnings()["not_exist"](path))
+            logging.warning(
                 "The file was loaded as a BioProv object, but it does not exist on the specified path."
             )
         return None