+++ bioprov/src/main.py
@@ -2,7 +2,7 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.18a"
+__version__ = "0.1.19"
 
 """
 
@@ -24,22 +24,31 @@ This class also contains functions to read and write objects in JSON and tab-del
 
 import datetime
 import json
-import pandas as pd
+import logging
 import tempfile
-from bioprov import config
-from bioprov.utils import Warnings, serializer, serializer_filter, dict_to_sha1
-from bioprov.src.files import File, SeqFile, Directory, deserialize_files_dict
-from bioprov.src.config import EnvProv
+from collections import OrderedDict
 from collections import deque
-from coolname import generate_slug
 from os import path
 from pathlib import Path
 from subprocess import Popen, PIPE, getoutput
 from time import time
 from types import GeneratorType
-from collections import OrderedDict
+
+import pandas as pd
+from coolname import generate_slug
 from prov.model import ProvEntity, ProvBundle, Namespace
 from tinydb import Query
+
+from bioprov import config
+from bioprov.src.config import EnvProv
+from bioprov.src.files import File, SeqFile, Directory, deserialize_files_dict
+from bioprov.utils import (
+    Warnings,
+    serializer,
+    serializer_filter,
+    dict_to_sha1,
+    create_logger,
+)
 
 
 class Program:
@@ -94,6 +103,10 @@ class Program:
             self._runs = dict()
         return self._runs
 
+    @runs.setter
+    def runs(self, value):
+        self._runs = value
+
     def add_runs(self, runs):
         """
         Sample method to add runs.
@@ -112,13 +125,14 @@ class Program:
         self.cmd = cmd
         return cmd
 
-    def add_parameter(self, parameter, _print=False):
+    def add_parameter(self, parameter, _generate_cmd=True):
         """
         Adds a parameter to the current instance and updates the command.
 
         :param parameter: an instance of the Parameter class.
-        :param _print: whether to print the parameter has been added.
-        :return:
+        :param _generate_cmd: Refreshes self.cmd when a Parameter is added.
+
+        :return: Updates self.params and self.cmd if _generate_cmd is True.
         """
         assert isinstance(parameter, Parameter), Warnings()["incorrect_type"](
             parameter, Parameter
@@ -126,23 +140,21 @@ class Program:
         k, v = parameter.key, parameter.value
         self.params[k] = parameter
         self.param_str = generate_param_str(self.params)
-        self.generate_cmd()
-        if _print:
-            print(
-                f"Added parameter {k} with value '{v}' to program {self.name}"
-            )  # no cover
+        config.logger.debug(
+            f"Added parameter {k} with value '{v}' to program {self.name}"
+        )  # no cover
+        if _generate_cmd:
+            self.generate_cmd()
 
-    def run(self, sample=None, _print=True):
+    def run(self, sample=None):
         """
         Runs the process.
         :param sample: An instance of the Sample class
-        :param _print: Argument to pass to Run.run()
         :return: An instance of Run class.
         """
-
         # Creates Run instance with self
         run_ = Run(self, sample=sample)
-        run_.run(_sample=sample, _print=True)
+        run_.run(_sample=sample)
         self.add_runs(run_)
         return run_
 
@@ -320,15 +332,12 @@ class Run:
         dict_ = {True: "Finished", False: "Pending"}
         return dict_[finished_status]
 
-    def run(self, _sample=None, _print=True, _print_stdout=False, _print_stderr=False):
+    def run(self, _sample=None):
         """
         Runs process for the Run instance.
         Will update attributes accordingly.
-        :type _print: bool
         :param _sample: self.sample
-        :param _print_stdout: Whether to print the stdout of the Program.
-        :param _print_stderr: Whether to print the stderr of the Program.
-        :return: self.stdout
+        :return: self
         """
         if _sample is None:
             _sample = self.sample
@@ -337,34 +346,35 @@ class Run:
         assert (
             self.program.found
         ), f"Cannot find program {self.program.name}. Make sure it is on your $PATH."
-        if _print:
-            str_ = f"Running program '{self.program.name}'"
-            if _sample is not None:
-                str_ += f" for sample {_sample.name}."
-            else:
-                str_ += "."
 
-            # Pretty printing of commands
-            split_ = self.program.cmd.split()
-            if len(self.program.cmd) > 80:
-                if len(split_) % 2 == 1:
-                    bin_, *fmt_cmd = split_
-                    last = ""
-                else:
-                    bin_, *fmt_cmd, last = split_  # no cover
-                it = iter(fmt_cmd)
-                fmt_cmd = zip(it, it)
-                fmt_cmd = " \\ \n".join(
-                    [bin_] + ["\t" + i[0] + " " + i[1] for i in fmt_cmd] + ["\t" + last]
-                )
-                str_ += f"\nCommand is:\n{fmt_cmd}\n"
-            else:
-                str_ += f"\nCommand is:\n{self.program.cmd}\n"
+        # Print block
+        str_ = f"Running program '{self.program.name}'"
+        if _sample is not None:
+            str_ += f" for sample {_sample.name}."
+        else:
+            str_ += "."
 
-            str_ = str_.strip()
-            if str_.endswith("\\"):
-                str_ = str_[:-1]
-            print(str_)
+        # Pretty printing of commands
+        split_ = self.program.cmd.split()
+        if len(self.program.cmd) > 80:
+            if len(split_) % 2 == 1:
+                bin_, *fmt_cmd = split_
+                last = ""
+            else:
+                bin_, *fmt_cmd, last = split_  # no cover
+            it = iter(fmt_cmd)
+            fmt_cmd = zip(it, it)
+            fmt_cmd = " \\ \n".join(
+                [bin_] + ["\t" + i[0] + " " + i[1] for i in fmt_cmd] + ["\t" + last]
+            )
+            str_ += f"\nCommand is:\n{fmt_cmd}\n"
+        else:
+            str_ += f"\nCommand is:\n{self.program.cmd}\n"
+
+        str_ = str_.strip()
+        if str_.endswith("\\"):
+            str_ = str_[:-1]
+        config.logger.info(str_)
 
         p = Popen(self.program.cmd, shell=True, stdout=PIPE, stderr=PIPE)
         self.process = p
@@ -387,11 +397,9 @@ class Run:
         self.finished = True
         self._status = self._finished_to_status(self.finished)
 
-        # These are useful for quick debugging.
-        if _print_stdout:
-            print(self.stdout)  # no cover
-        if _print_stderr:
-            print(self.stderr)  # no cover
+        if not self._auto_suppress_stdout:
+            config.logger.debug(self.stdout)  # no cover
+        config.logger.debug(self.stderr)  # no cover
 
         return self
 
@@ -476,7 +484,6 @@ class PresetProgram(Program):
         self.output_files = output_files
         self.preffix_tag = preffix_tag
         self.ready = False
-        self.generate_cmd()
 
         if self.sample is not None:
             self.create_func(sample=self.sample, preffix_tag=self.preffix_tag)
@@ -502,7 +509,7 @@ class PresetProgram(Program):
             param = Parameter(
                 key=k, value=str(self.sample.files[tag]), kind="input", tag=tag
             )
-            self.add_parameter(param)
+            self.add_parameter(param, _generate_cmd=False)
 
     def _parse_output_files(self):
         """
@@ -531,7 +538,7 @@ class PresetProgram(Program):
                 param = Parameter(
                     key=key, value=str(self.sample.files[tag]), kind="output", tag=tag
                 )
-                self.add_parameter(param, _print=False)
+                self.add_parameter(param, _generate_cmd=False)
         except ValueError:
             raise Exception(
                 "Please check the output files dictionary:\n'{}'\n"
@@ -559,8 +566,12 @@ class PresetProgram(Program):
         self._parse_input_files()
         self._parse_output_files()
 
+        # Add self to sample automatically
+        self.sample.add_programs(self)
+
         # Set ready state
         self.ready = True
+        self.generate_cmd()
 
     def validate_sample(self):
         """
@@ -578,13 +589,11 @@ class PresetProgram(Program):
         """
         assert isinstance(self, Program), Warnings()["incorrect_type"](self, Program)
 
-    def generate_cmd(self, from_files=True):
+    def generate_cmd(self):
         """
         TODO: improve this function
 
         Generates a wildcard command string, independent of samples.
-        :param from_files: Generate command from self.input_files and self.output_files (recommended) If False,
-        will generate from parameter dictionary instead.
         :return: Updates self.cmd.
         """
         self.validate_program()
@@ -597,7 +606,7 @@ class PresetProgram(Program):
                 try:
                     parameter.value = str(self.sample.files[f"{parameter.tag}"])
                 except AttributeError:
-                    print("Warning: no sample associated with program.")
+                    config.logger.warning("Warning: no sample associated with program.")
                     pass  # Suppress bug for now.
             else:
                 pass
@@ -608,11 +617,10 @@ class PresetProgram(Program):
         self.cmd = generic_cmd
         return generic_cmd
 
-    def run(self, sample=None, _print=True, preffix_tag=None):
+    def run(self, sample=None, preffix_tag=None):
         """
         Runs PresetProgram for sample.
         :param sample: Instance of bioprov.Sample.
-        :param _print: Whether to print more output.
         :param preffix_tag: Preffix tag to self.create_func()
         :return:
         """
@@ -620,11 +628,12 @@ class PresetProgram(Program):
             sample = self.sample
         if preffix_tag is None:
             preffix_tag = self.preffix_tag
-        if not self.ready:
-            self.create_func(sample, preffix_tag)
 
-        # Update self._run, run self.run() and update self._run again.
-        Program.run(self, sample=sample, _print=_print)
+        self.create_func(sample, preffix_tag)
+        run_ = Run(self, sample=sample)
+        run_.run(_sample=sample)
+        self.add_runs(run_)
+        return run_
 
 
 def parse_params(params):
@@ -676,7 +685,9 @@ def generate_param_str(params):
         param_str = str_.strip()
     else:
         # TODO: add more parameters options. List of tuples, List of Parameter instances, etc.
-        print("Must provide either a string or a dictionary for the parameters!")
+        config.logger.error(
+            "Must provide either a string or a dictionary for the parameters!"
+        )
         raise TypeError
     # Add positional arguments
     split_str = param_str.split()
@@ -717,6 +728,8 @@ def _add_programs(object_, programs):
     # Finally, append programs to object_.programs
     for program in programs:
         object_.programs[program.name] = program
+
+    object_.auto_update_db()
 
 
 def _add_runs(object_, runs):
@@ -792,10 +805,15 @@ class Sample:
             attributes = dict()
         self.attributes = attributes
         self.programs = OrderedDict()
+        self.project = None
 
         # This is an attribute used by the src.prov module
         self.namespace_preffix = f"samples:{self.name}"
         self.files_namespace_preffix = None
+
+    def auto_update_db(self):
+        if self.project is not None:
+            self.project.auto_update_db()
 
     def __repr__(self):
         str_ = f"Sample {self.name} with {len(self.files)} file(s)."
@@ -846,33 +864,31 @@ class Sample:
         Custom serializer for Sample class. Serializes runs, programs, and files attributes.
         :return:
         """
-        keys = ["files_namespace_preffix"]
+        keys = ["files_namespace_preffix", "project"]
         return serializer_filter(self, keys)
 
-    def run_programs(self, _print=True):
+    def run_programs(self):
         """
         Runs self._programs in order.
         :return:
         """
-        _run_programs(self, _print)
+        _run_programs(self)
 
-    def _run_program(self, program, _print=True):
+    def _run_program(self, program):
         """
         Runs program for self.
         :param program: bioprov.Program or bioprov.PresetProgram.
-        :param _print: Whether to print output of Program.
         :return: Runs Program and updates self.
         """
-        _run_program(self, program, _print)
+        _run_program(self, program)
 
-    def to_json(self, _path=None, _print=True):
+    def to_json(self, _path=None):
         """
         Exports the Sample as JSON. Similar to Project.to_json()
         :param _path: JSON output file path.
-        :param _print: Whether to print if the file was created correctly.
         :return:
         """
-        return to_json(self, self.serializer(), _path, _print=_print)
+        return to_json(self, self.serializer(), _path)
 
     def to_series(self):
         """
@@ -883,7 +899,7 @@ class Sample:
         series = {}
 
         # Can't apply serializer_filter here.
-        keys = ["files_namespace_preffix", "namespace_preffix", "_programs"]
+        keys = ["files_namespace_preffix", "namespace_preffix", "_programs", "project"]
         modified_dict = self.__dict__.copy()
         for key in keys:
             try:
@@ -903,34 +919,33 @@ class Sample:
         return pd.Series(series)
 
 
-def _run_program(_object, program, _print=True):
+def _run_program(_object, program):
     """
     Runs program for _object.
 
     :param _object: bioprov.Project or bioprov.Sample
     :param program: bioprov.Program or bioprov.PresetProgram.
-    :param _print: Whether to print output of Program.
     :return: Runs Program and updates _object.
     """
     if program not in _object.programs.keys():
         _object.add_programs(program)
-    program.run(sample=_object, _print=_print)
+    program.run(sample=_object)
+    _object.auto_update_db()
 
 
-def _run_programs(_object, _print=True):
+def _run_programs(_object):
     """
     Runs programs in order.
 
     :param _object: bioprov.Project or bioprov.Sample
-    :param _print: Whether to print output of Program.
     :return:
     """
     if len(_object.programs) >= 1:
         for _, p in _object.programs.items():
             # noinspection PyProtectedMember
-            _object._run_program(p, _print=_print)
+            _object._run_program(p)
     else:
-        print(f"No programs to run for {_object}")
+        config.logger.warning(f"No programs to run for {_object}")
 
 
 class Project:
@@ -938,7 +953,14 @@ class Project:
     Class which holds a dictionary of Sample instances, where each key is the sample name.
     """
 
-    def __init__(self, samples=None, tag=None, db=None, auto_update=False):
+    def __init__(
+        self,
+        tag=None,
+        samples=None,
+        db=None,
+        auto_update=False,
+        log_to_file=False,
+    ):
         """
         Initiates the object by creating a sample dictionary.
         :param samples: An iterator of Sample objects.
@@ -946,6 +968,8 @@ class Project:
         :param db: path to TinyDB to store project in JSON format.
         :param auto_update: Whether to auto_update the BioProvDB record.
                             Disabled by default.
+        :param log_to_file: Whether to log the Project to a File. You can define this later with
+                            the self.start_logging() method.
         """
         if tag is None:
             tag = generate_slug(2)
@@ -976,6 +1000,13 @@ class Project:
             db = config.db
         self.db = db
 
+        # Log attributes
+        self.log_to_file = log_to_file
+        self.log_file = None
+        self.logger = None
+        if self.log_to_file:
+            self.start_logging()
+
     def __len__(self):
         return len(self._samples)
 
@@ -987,9 +1018,10 @@ class Project:
             value = self._samples[item]
             return value
         except KeyError:
-            print(f"Sample {item} not in Project.\n")
-            print("Check the following keys:")
-            print(" ", "\n  ".join(self.keys))
+            config.logger.error(
+                f"Sample {item} not in Project.\n" f"Check the following keys:" " ",
+                "\n  ".join(self.keys),
+            )
 
     def __setitem__(self, key, value):
         self._samples[key] = value
@@ -1049,10 +1081,9 @@ class Project:
             db = self.db
         result, query = self.query_db(db)
         if result:
-            print(f"Updating project '{self.tag}' at {db.db_path}")
             db.update(self.serializer(), query.tag == self.tag)
         else:
-            print(f"Inserting new project '{self.tag}' in {db.db_path}")
+            config.logger.info(f"Inserting new project '{self.tag}' in {db.db_path}")
             db.insert(self.serializer())
 
     def auto_update_db(self):
@@ -1094,6 +1125,30 @@ class Project:
             for _, file in sample.files.items():
                 file.replace_path(old_terms, new, warnings)
 
+    def start_logging(
+        self, log_file=None, level=logging.INFO, _custom_start_message=None
+    ):
+        """
+        Starts logging Project to File.
+
+        :param log_file: Path to log file. If None will be defined automatically.
+        :param level: Logging level.
+        :param _custom_start_message: Custom starting message to start the log.
+        :return: Creates logger attributes and refreshes bp.config.logger
+        """
+        if log_file is None:
+            log_file = f"{self.tag}.log"
+        self.log_file = log_file
+        self.logger = config.logger = create_logger(level, self.log_file, self.tag)
+        if _custom_start_message is None:
+            _custom_start_message = f"Starting log for project '{self.tag}'."
+
+        self.logger.info(_custom_start_message)
+
+        self.add_files(File(self.log_file, tag="log"))
+        self.logger.info(f"Writing log to {self.files['log']}")
+        self.logger.info(f"Loading {len(self)} samples.")
+
     def add_files(self, files):
         """
         Adds Files to self.files. See documentation to bioprov.src.main.add_files().
@@ -1102,7 +1157,6 @@ class Project:
         :return: Updates self.files
         """
         _add_files(self, files)
-        self.auto_update_db()
 
     def add_programs(self, programs):
         """
@@ -1111,25 +1165,23 @@ class Project:
         :return: Updates self.programs
         """
         _add_programs(self, programs)
-        self.auto_update_db()
 
-    def run_programs(self, _print=True):
+    def run_programs(self):
         """
         Runs all programs in self.programs in order.
 
         :return:
         """
-        _run_programs(self, _print)
+        _run_programs(self)
 
-    def _run_program(self, program=None, _print=True):
+    def _run_program(self, program=None):
         """
         Runs program for self.
 
         :param program: bioprov.Program or bioprov.PresetProgram.
-        :param _print: Whether to print output of Program.
         :return: Runs Program and updates self.
         """
-        _run_program(self, program, _print)
+        _run_program(self, program)
 
     @staticmethod
     def is_sample_and_name(sample):
@@ -1146,7 +1198,9 @@ class Project:
         if sample.name is None:
             slug = generate_slug(2)
             sample.name = slug
-            print(f"No sample name set. Setting random name: {sample.name}")
+            config.logger.warning(
+                f"No sample name set. Setting random name: {sample.name}"
+            )
 
         return sample
 
@@ -1167,8 +1221,7 @@ class Project:
 
         return constructor
 
-    @staticmethod
-    def build_sample_dict(constructor):
+    def build_sample_dict(self, constructor):
         """
         Build sample dictionary from passed constructor.
         :param constructor: Iterable or NoneType
@@ -1184,6 +1237,7 @@ class Project:
         for sample in constructor:
             sample = Project.is_sample_and_name(sample)
             samples[sample.name] = sample
+            samples[sample.name].project = self
 
         return samples
 
@@ -1198,14 +1252,13 @@ class Project:
     def serializer(self):
         return serializer(self)
 
-    def to_json(self, _path=None, _print=True):
+    def to_json(self, _path=None):
         """
         Exports the Project as JSON. Similar to Sample.to_json()
         :param _path: JSON output file _path.
-        :param _print: Whether to print if the file was created correctly.
         :return:
         """
-        return to_json(self, self.serializer(), _path, _print=_print)
+        return to_json(self, self.serializer(), _path)
 
     def to_df(self):
         """
@@ -1291,11 +1344,13 @@ def _add_files(object_, files):
     # Here 'files' must be a dictionary of File or Directory instances
     for k, v in files.items():
         if k in object_.files.keys():
-            print(f"Updating file {k} with value {v}.")
+            config.logger.info(f"Updating file {k} with value {v}.")
         object_.files[k] = v
 
+    object_.auto_update_db()
 
-def to_json(object_, dictionary, _path=None, _print=True):
+
+def to_json(object_, dictionary, _path=None):
     """
     Exports the Sample or Project as JSON.
     :return: Writes JSON output
@@ -1307,7 +1362,7 @@ def to_json(object_, dictionary, _path=None, _print=True):
     if "json" not in object_.files.keys():
         object_.add_files({"json": _path})
 
-    return write_json(dictionary, _path, _print=_print)
+    return write_json(dictionary, _path)
 
 
 def from_json(json_file, kind="Project", replace_path=None, replace_home=False):
@@ -1363,7 +1418,7 @@ def from_json(json_file, kind="Project", replace_path=None, replace_home=False):
             pass
 
         # Create Project
-        project = Project(samples=samples, tag=d["tag"])
+        project = Project(tag=d["tag"], samples=samples)
 
         # Deserializing and adding project files and programs
         try:
@@ -1403,9 +1458,11 @@ def from_json(json_file, kind="Project", replace_path=None, replace_home=False):
             project.replace_paths(other_HOME_variables, HOME, warnings=True)
 
         if replace_path:
-            print("Replacing paths:")
-            print(f"\tOld:\t{replace_path[0][0]}")
-            print(f"\tNew:\t{replace_path[1]}")
+            config.logger.info(
+                "Replacing paths:"
+                f"\tOld:\t{replace_path[0][0]}"
+                f"\tNew:\t{replace_path[1]}"
+            )
             project.replace_paths(replace_path[0], replace_path[1], warnings=True)
 
         return project
@@ -1477,7 +1534,7 @@ def from_df(
                 sample.attributes[attr_] = row[attr_]
         samples[ix] = sample
 
-    samples = Project(samples, tag=tag)
+    samples = Project(tag=tag, samples=samples)
     if source_file:
         samples.add_files({"project_csv": source_file})
 
@@ -1545,22 +1602,20 @@ def dict_to_sample(json_dict):
     return sample_
 
 
-def write_json(dict_, _path, _print=True):
+def write_json(dict_, _path):
     """
     Writes dictionary to JSON file.
     :param dict_: JSON dictionary.
     :param _path: String with _path to JSON file.
-    :param _print: Whether to print if the file was successfully created.
     :return: Writes JSON file.
     """
     with open(_path, "w") as f:
         json.dump(dict_, f, indent=3)
 
-    if _print:
-        if Path(_path).exists():
-            print(f"Created JSON file at {_path}.")
-        else:
-            print(f"Could not create JSON file for {_path}.")
+    if Path(_path).exists():
+        config.logger.info(f"Created JSON file at {_path}.")
+    else:
+        config.logger.info(f"Could not create JSON file for {_path}.")
 
 
 def load_project(tag):
@@ -1578,7 +1633,7 @@ def load_project(tag):
     try:
         result = config.db.search(query.tag == tag)[0]
     except (IndexError, KeyError):
-        print(f"Project not found in database at {config.db_path}")
+        config.logger.error(f"Project not found in database at {config.db_path}")
         return
 
     with tempfile.NamedTemporaryFile() as f: