+++ bioprov/tests/test_bioprov_workflows.py
@@ -2,15 +2,16 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.18a"
+__version__ = "0.1.19"
 
 
 """
 Testing for the workflows package.
 """
 from os import remove
-from bioprov.utils import Warnings
+from pathlib import Path
 from bioprov.data import genome_annotation_dataset
+from bioprov.utils import Warnings
 from bioprov.workflows.blastn import blastn_alignment
 from bioprov.workflows.genome_annotation import genome_annotation
 from bioprov.workflows.kaiju import KaijuWorkflow
@@ -32,7 +33,7 @@ def test_genome_annotation():
     Tests the 'genome_annotation' workflow with the 'prodigal' step.
     :return:
     """
-    workflow = genome_annotation()
+    workflow = genome_annotation(tag="test-project")
     workflow.input = genome_annotation_dataset
     steps = [
         "prodigal",
@@ -42,6 +43,10 @@ def test_genome_annotation():
     for _, sample in workflow.project.items():
         for key, file in sample.files.items():
             assert file.exists, Warnings()["not_exist"](file.path)
+
+    log_file = "test-project.log"
+    assert Path(log_file).exists()
+    remove(log_file)
 
 
 def test_kaiju_workflow():