+++ bioprov/workflows/__init__.py
@@ -2,10 +2,10 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.18a"
+__version__ = "0.1.19"
 
 
+from .blastn import blastn_alignment
 from .genome_annotation import genome_annotation
 from .kaiju import KaijuWorkflow
-from .blastn import blastn_alignment
 from .wf_parser import WorkflowOptionsParser